+++ packages/schemas/tsconfig.json
@@ -1,27 +1,6 @@
 {
+  "extends": "./tsconfig.esm",
   "compilerOptions": {
-    "noImplicitAny": true,
-    "target": "ES2015",
-    "jsx": "react",
-    "esModuleInterop": true,
-    "isolatedModules": true,
-    "declaration": true,
-    "declarationDir": "dist/types",
-    "module": "ES2015",
-    "lib": ["es6", "dom", "es2016", "es2017"],
-    "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "types": ["node", "jest", "@types/jest"],
-    "typeRoots": ["../node_modules/@types"],
-    "sourceMap": true,
-    "baseUrl": "..", 
-    "paths": {
-      "@pdfme/common": ["packages/common/src"], 
-      "@pdfme/schemas": ["packages/schemas/src"] 
-    }
-  },
-  "baseUrl": ".",
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["node_modules"]
+    "outDir": "./dist",
+  }
 }
